requirements.txt
albumentations==1.0.0
effdet==0.2.4
ensemble-boxes==1.0.6
torchmetrics==0.6.0
fastcore==1.3.20
timm==0.4.9
git+https://github.com/alexhock/object-detection-metrics
pytorch-lightning==1.3.5
setuptools==59.5.0
pip install torch==1.11.0+cu113 torchvision==0.12.0+cu113 torchaudio==0.11.0 --extra-index-url https://download.pytorch.org/whl/cu113
